minor stylistic improvements to /view/
Before now, short files would have a really short content block. Now, I set a minimum height for said block.

Before now, the mobile version would have the code block with a #f0f0f0 background, then a body under it with a background #d5d5d5. This looked really bad, so now the body on the mobile version matches the #f0f0f0 of the content to make it look as though the content extends to the bottom of the screen.