don't let the browser cache the upload/list pages
For the list page, the reasoning behind this is obvious: whenever the content on the site changes, the contents of the list page changes. For the upload page, the reason is a bit more subtle. When login/logout is supported, the user might be logged out and then navigate to the homepage; if they do this, the homepage should not be the upload page, but rather a redirect to the login page. If their browser already cached the upload page, then they'd see an upload page which would inevitably fail to post.